adaptive-pipeline-model-function/requirements.txt
google-auth
google-cloud-batch
google-api-core